more missing mapped attributes added

+++ app/adapter/in/graphql/graph/model/models_gen.go
@@ -324,6 +324,6 @@ type Vehicle struct {
 }
 
 type Weight struct {
-	Unit  *string `json:"unit,omitempty"`
-	Value *int    `json:"value,omitempty"`
+	Unit  *string  `json:"unit,omitempty"`
+	Value *float64 `json:"value,omitempty"`
 }
